requirements.txt
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
albumentations>=1.0.3
matplotlib>=3.4.3
tensorboard>=2.6.0
tqdm>=4.62.2
scikit-learn>=0.24.2
pydicom>=2.2.2
SimpleITK>=2.1.1
flask>=2.0.1
transformers>=4.11.0
pillow>=8.3.1
streamlit>=1.0.0
onnx>=1.10.0
onnxruntime>=1.9.0
pandas>=1.3.3
scikit-image>=0.18.3 